Improvements and ready to implement music rooms

+++ music_individual/src/main/java/music_individual/demo/business/impl/UsersManagerImpl.java
@@ -2,9 +2,12 @@ package music_individual.demo.business.impl;
 
 import lombok.AllArgsConstructor;
 import music_individual.demo.business.IUsersManager;
-import music_individual.demo.domain.UpdateUserRequest;
+import music_individual.demo.business.exception.FailedCRUDException;
+import music_individual.demo.business.exception.ObjectMissingException;
 import music_individual.demo.persistence.UsersRepusitory;
 import music_individual.demo.persistence.entities.UserEntity;
+import org.hibernate.ObjectDeletedException;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -26,37 +29,43 @@ public class UsersManagerImpl implements IUsersManager {
     }
 
     @Override
-    public void AddUser(UserEntity user){
+    public UserEntity AddUser(UserEntity user){
 
         String encodedPassword = passwordEncoder.encode(user.getPassword());
         user.setPassword(encodedPassword);
-        repo.save(user);
+        return repo.save(user);
     }
 
     @Override
-    public void UpdateUser(UpdateUserRequest request){
-        Optional<UserEntity> userOptional = repo.findById(request.getId());
+    public void UpdateUser(UserEntity user){
+        Optional<UserEntity> userOptional = repo.findById(user.getId());
         if(userOptional.isPresent()) {
-            UserEntity user = userOptional.get();
+            UserEntity updatedUser = userOptional.get();
 
-
-            user.setEmail(request.getEmail());
-            user.setPassword(request.getPassword());
-            repo.save(user);
+            updatedUser.setEmail(user.getEmail());
+            updatedUser.setPassword(user.getPassword());
+            repo.save(updatedUser);
+        }
+        else{
+            throw new ObjectMissingException();
         }
 
     }
 
     @Override
-    public void DeleteUser(Integer id){
-
+    public boolean DeleteUser(Integer id){
+        if(repo.findById(id).isPresent()){
+            throw new FailedCRUDException("User was not deleted successfully");
+        }
         repo.deleteById(id);
+        return true;
     }
 
     @Override
     public UserEntity GetUserByID(Integer id) {
-
-
+        if(!repo.findById(id).isPresent()){
+            throw new ObjectMissingException();
+        }
         return repo.findById(id).get();
     }
 }
